Renames for files and imports for jest-haste-map
