+++ packages/jest-haste-map/src/getMockName.js
