+++ packages/jest-haste-map/src/lib/WatchmanWatcher.js
