+++ packages/jest-haste-map/src/lib/__tests__/normalizePathSep.test.js
@@ -13,14 +13,14 @@ describe('normalizePathSep', () => {
   it('does nothing on posix', () => {
     jest.resetModules();
     jest.mock('path', () => jest.requireActual('path').posix);
-    const normalizePathSep = require('../normalize_path_sep').default;
+    const normalizePathSep = require('../normalizePathSep').default;
     expect(normalizePathSep('foo/bar/baz.js')).toEqual('foo/bar/baz.js');
   });
 
   it('replace slashes on windows', () => {
     jest.resetModules();
     jest.mock('path', () => jest.requireActual('path').win32);
-    const normalizePathSep = require('../normalize_path_sep').default;
+    const normalizePathSep = require('../normalizePathSep').default;
     expect(normalizePathSep('foo/bar/baz.js')).toEqual('foo\\bar\\baz.js');
   });
 });
