+++ packages/jest-haste-map/src/lib/extractRequires.js
