+++ packages/jest-haste-map/src/lib/getPlatformExtension.js
