+++ packages/jest-haste-map/src/lib/normalizePathSep.js
